+++ xcm_integration_tests/emulated/chains/parachains/frontier-template-dancelight/Cargo.toml
@@ -12,9 +12,10 @@ workspace = true
 workspace = true
 
 [dependencies]
-container-chain-template-simple-runtime  = { workspace = true, features = [ "std" ] }
 emulated-integration-tests-common = { workspace = true }
 sp-core = { workspace = true }
-tanssi-emulated-integration-tests-common = { workspace = true }
 xcm = { workspace = true }
 xcm-emulator = { workspace = true }
+container-chain-template-frontier-runtime = { workspace = true }
+cumulus-primitives-core = { workspace = true }
+hex-literal = { workspace = true }
