+++ xcm_integration_tests/emulated/chains/parachains/frontier-template-dancelight/src/genesis.rs
@@ -14,40 +14,56 @@
 // You should have received a copy of the GNU General Public License
 // along with Tanssi.  If not, see <http://www.gnu.org/licenses/>.
 
-use {
-    container_chain_template_simple_runtime::UNIT as DEV,
-    emulated_integration_tests_common::build_genesis_storage,
-    tanssi_emulated_integration_tests_common::accounts,
-};
+use container_chain_template_frontier_runtime::AccountId;
+use {emulated_integration_tests_common::build_genesis_storage, hex_literal::hex};
 
 pub const PARA_ID: u32 = 2002;
 pub const ORCHESTRATOR: u32 = 2000;
-const ENDOWMENT: u128 = 1_000_000 * DEV;
 
 pub fn genesis() -> sp_core::storage::Storage {
-    let genesis_config = container_chain_template_simple_runtime::RuntimeGenesisConfig {
-        balances: container_chain_template_simple_runtime::BalancesConfig {
-            balances: accounts::init_balances()
+    let genesis_config = container_chain_template_frontier_runtime::RuntimeGenesisConfig {
+        system: Default::default(),
+        balances: container_chain_template_frontier_runtime::BalancesConfig {
+            balances: pre_funded_accounts()
                 .iter()
                 .cloned()
-                .map(|k| (k, ENDOWMENT))
+                .map(|k| (k, 1 << 80))
                 .collect(),
         },
-        parachain_info: container_chain_template_simple_runtime::ParachainInfoConfig {
+        parachain_info: container_chain_template_frontier_runtime::ParachainInfoConfig {
             parachain_id: PARA_ID.into(),
             ..Default::default()
         },
-        sudo: container_chain_template_simple_runtime::SudoConfig {
-            key: Some(accounts::init_balances()[0].clone()),
+        // EVM compatibility
+        // We should change this to something different than Moonbeam
+        // For now moonwall is very tailored for moonbeam so we need it for tests
+        evm_chain_id: container_chain_template_frontier_runtime::EVMChainIdConfig {
+            chain_id: 1281,
+            ..Default::default()
         },
-        authorities_noting: container_chain_template_simple_runtime::AuthoritiesNotingConfig {
+        sudo: container_chain_template_frontier_runtime::SudoConfig {
+            key: Some(pre_funded_accounts()[0]),
+        },
+        authorities_noting: container_chain_template_frontier_runtime::AuthoritiesNotingConfig {
             orchestrator_para_id: ORCHESTRATOR.into(),
             ..Default::default()
         },
         ..Default::default()
     };
+
     build_genesis_storage(
         &genesis_config,
-        container_chain_template_simple_runtime::WASM_BINARY.unwrap(),
+        container_chain_template_frontier_runtime::WASM_BINARY.unwrap(),
     )
+}
+/// Get pre-funded accounts
+pub fn pre_funded_accounts() -> Vec<AccountId> {
+    // These addresses are derived from Substrate's canonical mnemonic:
+    // bottom drive obey lake curtain smoke basket hold race lonely fit walk
+    vec![
+        AccountId::from(hex!("f24FF3a9CF04c71Dbc94D0b566f7A27B94566cac")), // Alith
+        AccountId::from(hex!("3Cd0A705a2DC65e5b1E1205896BaA2be8A07c6e0")), // Baltathar
+        AccountId::from(hex!("798d4Ba9baf0064Ec19eB4F0a1a45785ae9D6DFc")), // Charleth
+        AccountId::from(hex!("773539d4Ac0e786233D90A233654ccEE26a613D9")), // Dorothy
+    ]
 }
