+++ xcm_integration_tests/emulated/chains/parachains/frontier-template-dancelight/src/lib.rs
@@ -15,4 +15,29 @@
 // along with Tanssi.  If not, see <http://www.gnu.org/licenses/>.
 
 pub mod genesis;
+use xcm_emulator::decl_test_parachains;
 pub use xcm_emulator::{bx, TestExt};
+
+decl_test_parachains! {
+    // Dancelight parachains
+    pub struct FrontierTemplateDancelight {
+        genesis = genesis::genesis(),
+        on_init = (),
+        runtime = container_chain_template_frontier_runtime,
+        core = {
+            XcmpMessageHandler: container_chain_template_frontier_runtime::XcmpQueue,
+            LocationToAccountId: container_chain_template_frontier_runtime::xcm_config::LocationToAccountId,
+            ParachainInfo: container_chain_template_frontier_runtime::ParachainInfo,
+            MessageOrigin: cumulus_primitives_core::AggregateMessageOrigin,
+        },
+        pallets = {
+            System: container_chain_template_frontier_runtime::System,
+            Balances: container_chain_template_frontier_runtime::Balances,
+            ParachainSystem: container_chain_template_frontier_runtime::ParachainSystem,
+            PolkadotXcm: container_chain_template_frontier_runtime::PolkadotXcm,
+            ForeignAssets:  container_chain_template_frontier_runtime::ForeignAssets,
+            AssetRate:  container_chain_template_frontier_runtime::AssetRate,
+            ForeignAssetsCreator: container_chain_template_frontier_runtime::ForeignAssetsCreator,
+        }
+    },
+}
