+++ xcm_integration_tests/emulated/chains/parachains/simple-template-dancelight/src/lib.rs
@@ -16,3 +16,27 @@
 
 pub mod genesis;
 pub use xcm_emulator::{bx, TestExt};
+use {frame_support::parameter_types, xcm_emulator::decl_test_parachains};
+
+decl_test_parachains! {
+    pub struct SimpleTemplateDancelight {
+        genesis = genesis::genesis(),
+        on_init = (),
+        runtime = container_chain_template_simple_runtime,
+        core = {
+            XcmpMessageHandler: container_chain_template_simple_runtime::XcmpQueue,
+            LocationToAccountId: container_chain_template_simple_runtime::xcm_config::LocationToAccountId,
+            ParachainInfo: container_chain_template_simple_runtime::ParachainInfo,
+            MessageOrigin: cumulus_primitives_core::AggregateMessageOrigin,
+        },
+        pallets = {
+            System: container_chain_template_simple_runtime::System,
+            Balances: container_chain_template_simple_runtime::Balances,
+            ParachainSystem: container_chain_template_simple_runtime::ParachainSystem,
+            PolkadotXcm: container_chain_template_simple_runtime::PolkadotXcm,
+            ForeignAssets:  container_chain_template_simple_runtime::ForeignAssets,
+            AssetRate:  container_chain_template_simple_runtime::AssetRate,
+            ForeignAssetsCreator: container_chain_template_simple_runtime::ForeignAssetsCreator,
+        }
+    }
+}
